split story carousel components
